Refactor LLMTestGenerator: implement new prompt generation and execution methods, remove legacy test_generator
